Add examples for advanced features in DynamicList
- Introduced AdvancedFeaturesExample.swift showcasing optional detail content and default views.
- Added BasicListsExample.swift demonstrating static data lists and user actions.
- Implemented ReactiveListsExample.swift to illustrate reactive data handling with publishers and error management.
- Created SectionedListsExample.swift to display sectioned lists with search capabilities.
- Added SharedExampleData.swift for shared models and sample data across examples.
- Removed outdated examples including CustomActionsExample.swift, DefaultViewsPreviews.swift, and others to streamline the codebase.
